extract module.go and constants.go from yarpcfx
